fix: removing crypto dependency
