+++ package.json
@@ -77,7 +77,7 @@
     "@oaknational/oak-components": "^1.47.0",
     "@oaknational/oak-consent-client": "^2.1.1",
     "@oaknational/oak-curriculum-schema": "^1.36.0",
-    "@oaknational/oak-pupil-client": "^2.12.1",
+    "@oaknational/oak-pupil-client": "^2.12.3",
     "@portabletext/react": "^3.0.11",
     "@react-aria/aria-modal-polyfill": "^3.7.8",
     "@sanity/asset-utils": "^1.3.0",
@@ -121,6 +121,7 @@
     "stream-http": "^3.2.0",
     "styled-components": "^5.3.6",
     "swr": "^2.2.4",
+    "tiny-hashes": "^1.0.1",
     "url": "^0.11.2",
     "util": "^0.12.5",
     "uuid": "^11.0.2",
