+++ src/pages/teachers/curriculum/index.tsx
@@ -73,9 +73,9 @@ const CurriculumHomePage: NextPage<CurriculumHomePageProps> = (props) => {
             <Hr $color={"white"} $mb={0} />
           </Box>
           <Flex $mt={[24, 80]} $mb={[80]}>
-            <HomepageCurriculumLandingHero
+            {/* <HomepageCurriculumLandingHero
               subjectPhaseOptions={subjectPhaseOptions}
-            />
+            /> */}
           </Flex>
         </MaxWidth>
       </Flex>
@@ -203,9 +203,7 @@ const CurriculumHomePage: NextPage<CurriculumHomePageProps> = (props) => {
   );
 };
 
-export const filterValidSubjectPhaseOptions = (
-  subjects: SubjectPhaseOption[],
-) => {
+const filterValidSubjectPhaseOptions = (subjects: SubjectPhaseOption[]) => {
   subjects.forEach(({ ks4_options }) => {
     if (
       ks4_options &&
@@ -222,7 +220,7 @@ export const filterValidSubjectPhaseOptions = (
   return subjects;
 };
 
-export const fetchSubjectPhasePickerData: () => Promise<SubjectPhasePickerData> =
+const fetchSubjectPhasePickerData: () => Promise<SubjectPhasePickerData> =
   async () => {
     const subjects = await curriculumApi2023.subjectPhaseOptions({
       cycle: "2",
