+++ src/types/tiny-hashes.d.ts
@@ -1,0 +1,3 @@
+declare module "tiny-hashes/sha256" {
+  export default function sha256(input: string): string;
+}
